Make MasterBff its own generator, and move the code to its own file.
